requirements.txt
streamlit
llama_index
python-dotenv
llama-index-llms-groq
trafilatura
